use biggest avatar for v3 api
